+++ internal/cli/add.go
@@ -42,6 +42,8 @@ func init() {
 	addCmd.Flags().BoolVar(&addEncrypt, "encrypt", false, "encrypt file when adding")
 	addCmd.Flags().StringVarP(&addSource, "source", "s", "", "target source directory")
 	addCmd.Flags().BoolVar(&addTemplate, "template", false, "mark file as template")
+
+	addCmd.RegisterFlagCompletionFunc("source", completeSources)
 }
 
 func runAdd(cmd *cobra.Command, args []string) error {